Refactor scripting module to remove dependency to AFS
