+++ loadflow/pom.xml
@@ -26,6 +26,7 @@
     <modules>
         <module>loadflow-api</module>
         <module>loadflow-results-completion</module>
+        <module>loadflow-scripting</module>
         <module>loadflow-validation</module>
     </modules>
