--- scripting/scripting-afs/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,3 +1,0 @@
-moduleName = powsybl-scripting-module
-moduleVersion = 1.0
-extensionClasses = com.powsybl.scripting.groovy.AfsGroovyExtensionModule
