--- scripting/scripting-extension/src/main/java/com/powsybl/scripting/groovy/GroovyScriptExtension.java
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) 2016, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.scripting.groovy;
-
-import com.powsybl.computation.ComputationManager;
-import groovy.lang.Binding;
-
-/**
- * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
- */
-public interface GroovyScriptExtension {
-
-    void load(Binding binding, ComputationManager computationManager);
-}
