+++ scripting/src/main/groovy/com/powsybl/scripting/groovy/InitPowsybl.groovy
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2018, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.scripting.groovy
+
+import org.codehaus.groovy.tools.shell.CommandSupport
+import org.codehaus.groovy.tools.shell.Groovysh
+
+/**
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ */
+class InitPowsybl extends CommandSupport {
+
+    protected InitPowsybl(final Groovysh shell) {
+        super(shell, ':init_powsybl', ':powsybl')
+    }
+
+    Object execute(List args) {
+    }
+}
