+++ scripting/src/main/resources/com/powsybl/scripting/groovy/InitPowsybl.properties
